Updated download format code in split.py
To address issue #55, discovered that  when we download the room-long stream, it comes down in 720p, despite us requesting best video/best audio.   The culprit was a conditional where the container_format automatically set the download_format to "mp4"

Removing the conditional was the main fix, but took the opportunity to update the download format using the example from yt-dlp manual, to ensure that we are downloading in mp4.